Moved back options settings init for ACD due to fatal error, changed WP Rocket compatibility so that it works with ACD/FPC
